Cleaned up AST; e.g. moved functionality to better places
- 'Node' no longer has children, as only BlockNode uses it. There-
fore the 'add_child' method is no longer required, and the equals
and __str__ function of 'Node' are now shorter.
- Fixed indentation for IfElse block.
- Added operator field to MonOps and BinOps as they are needed in
the TypeChecking phase and by adding them to the AST it is no
longer required to manually add them in function calls, which
seemed messy to me.
- Removed 'accept' methods from expression parent classes, as they
are never called.
- Added equals and print methods to expression parent classes, as
they are always the same for the child expression classes.
- Added a TypeNode, which also gets the equals and print methods
for the same reason.