Start of a LARGE commit: Parser.py and AST.py
Parser.py:
The most important thing that forced me to change is that I
renamed the Type nodes into Var nodes, and created a new series
of Type nodes that describe the 'type names', as they appear in
questions, such as 'integer', 'boolean'.

So, now instead of just a string, each question has an XyzTypeNode.
This allowed me to remove the TypeEnums.py file, which I, in the
end, figured was a bad idea after all. Tijs talked about the
barrier between before-AST and after-AST, and that class was being
used in both phases, so I figured it had to go. More on the AST
changes in a later commit.

Some smaller things that happened in Parser.py:
- I changed the abbreviated indexes, e.g. 'BOOL' and 'INT', to
their full-length counterparts, for consistency.
- Shortened the DECIMAL and MONEY regular expressions.
- Some re-ordering of functions that seemed better to me, and some
newlines added to comply to PEP8.

AST.py:
As said above, I added the 'XyzTypeNode' classes, which has the
TypeNode as parent class. I figured it was best to call the
classes that contain variables and literals children of 'VarNode'.
Not entirely sure if that really fits, but eh... Either way, the
VarNode thus now became a parent node.

The new TypeNode features a lot of the contents that TypeEnums.py
first had. It features additional features to easily check whether
a type is e.g. a boolean, numeric, alphanumeric. This helps making
the TypeChecking phase more clearer to 'read' and understand. More
on that later.