Part #2: TypeChecker, Environment, Evaluator, ErrorHandler
TypeChecker.py:
The typechecker was quite heavily dependant on the TypeEnums class.
So, I had to change it quite a bit to make it work. The biggest
changes is that the 'general' functions to determine whether minOp
and binOps feature correct types is now gone.

It is best to start all the way down; the bottom 'xyz_node'
functions now no longer return a string that indicates the type of
the literal/variable, but it returns a XyzTypeNode as I newly
created in the AST. This helps tremendously, because these nodes
feature methods to check whether they are a certain type or part
of a certain group.

Let's go back up. In 'highest_number_type', you can see that 'left'
is a TypeNode, that features the 'is_decimal' function. This is an
example of the additional functionality that was added by adding
these nodes to the AST. This is then further applied to the whole
TypeChecking file.

In each of the functions like 'mon_op_node' and 'add_node', the
generic call of a validation function has been replaced by logic-
like expressions. Because of the new functionality, the conditions
on which the types are valid can be spelled out in more detail in
the functions themselves. This makes it much easier to see what is
acceptable for certain nodes, and furthermore the use of the
'is_alphanumeric' functions seems better readable cleaner.

NOTE: the '^' operator works as XOR on booleans.

Environment.py:
Quite a lot changed here, to support the changes featured in the
other files mentioned so far. For clarity, I changed some function
names; for example, add_var seemed more descriptive than get_type.

The default values that were featured here were moved to the newly
created 'XyzTypeNode' AST nodes. This is another point at which we
no longer need to index a dictionary. Furthermore, I merged the
two environments into one, and added the question node as a whole
instead of just the type. This is required for drawing the GUI.

This new environment dictionary is also of type 'OrderedDict', as
the chronological order of the questions is also required for
drawing the GUI correctly.

Evaluate.py:
In the evaluation phase, I added the error handler to be able to
add the warning about division by zero. In case it occurs, I just
return zero.

ErrorHandler.py:
Added an 'add_warning' function. This is better than having one
function that does both, as it adds a function path that leads
to an assertion. I think it's also more clearer. But yeah, I
probably would've done this as well ;)

Of course I added the division by zero warning function here. Some
general formatting corrections to finish it off.