Part #3: Fixing unitTests and adding some stuff to the main script
UnitTests/testParsing.py:
Sadly, because of all the stuff that I did, the unitTests stopped
working. In the end, the fix wasn't that hard, just change the
examples in which questions were defined. The 'q_type' variables
had to be changed into 'XyzTypeNode' AST nodes.

Unfortunately, I created some bugs in the Parser.py code, which
made debugging it pretty hard. In the end I reset some things in
the Parser.py code, after which the unitTests worked again. Yay!

StartQuestionnaire.py:
Well, Git sees this a new file, because I renamed it. There was a
spelling error in the name, so that's why. I slightly editted some
of the code that was already there, and commented the FindCycles
lines, as they prevented the program from parsing.

The main thing that really happened here was adding the call to
DrawGUI. The 'start_traversal' might need to be changed later ;p
Next commit will feature graphic stuff ;)