Part #4: Yay, we can draw the GUI!
DrawGUI.py:
This is the class that makes use of the environment to draw the
questions to the AppJar GUI. For each question in the environment,
the Type is visited to retrieve the GUI function that is default
for that type. Then, the information from the QuestionNode is used
to actually draw the question onscreen.

Here we can see the initial goal of the 'XyzTypeNode' classes. Now
we can select a standard GUI element without having to retrieve it
from some dictionary. It's again the visitor pattern.

FormGUI.py:
The code here was improved to be shorter, and to include the gene-
ration of more element types. It's not yet finished, but the form
in the repository fully works!

As can be seen, the defaults array has been deleted, as explained
above. Some of the elements that can now be drawn onscreen requi-
red some hacky stuff to get it to work. Using AppJar source code
and Tkinter calls, for example.

The actual redrawing is not yet added, but I don't think it is
that hard to add. After all the work on the other stuff, I am now
glad the first GUI with questions can be displayed!