We now have input changed events in the renderer.
There is still quite a code smell here in the switch case. Strings do
not work either because we do not have a StringValue class but I'm not
sure if this is by design or for some other reason. Anyway I committed
the refactor for value as well because it was disgusting.