Added a Undefined variable and let it be properly evaluated. You could say you can override all operators to return a Undefined object, however then you have trouble identifying the Undefined variable later on as all equality operators will return Undefined. Instanceof would be the solution we are now always trying to avoid.
Each field value in the form will be undefined untill a field
identifier gets a new value assigned. All questions types stay
undefined till a value is entered into the field, with the
exception of the boolean type. This one is always evaluated to
false as it would be strange to have a undefined value for a
checkbox.

At last, fixed a small bug in the FindCycles traversal. And silenced
some pycharm warnings in the AST.